Split Lsh functions into separate files. Rename LshFoo -> HashingFoo.
